requirements.txt
PySide6>=6.6.0
google-generativeai>=0.5.0
python-docx>=1.1.0
